+++ pkg/acs/lib/spline.c
@@ -34,7 +34,7 @@ int splint_nr (double *xa, double *ya, int n,
 	spline_nr (xa, ya, y2a, ua, n);
 
 	/* Evaluate the fit at each element of x. */
-	for (size_t i = 0;  i < nelem;  i++) {
+	for (int i = 0;  i < nelem;  i++) {
 
 	    /* Find the segment containing x[i]. */
 	    klo = 0;
